+++ examples/swr-site/theme.config.tsx
@@ -1,7 +1,7 @@
 /* eslint sort-keys: error */
 import { useRouter } from 'next/router'
 import type { DocsThemeConfig } from 'nextra-theme-docs'
-import { useConfig } from 'nextra-theme-docs'
+import { LocaleSwitch, useConfig } from 'nextra-theme-docs'
 import type { ComponentProps, ReactElement } from 'react'
 
 const SWRLogo = (props: ComponentProps<'svg'>): ReactElement => (
@@ -166,6 +166,9 @@ const config: DocsThemeConfig = {
       </>
     )
   },
+  navbar: {
+    extraContent: LocaleSwitch
+  },
   nextThemes: {
     defaultTheme: 'dark'
   },
@@ -187,6 +190,7 @@ const config: DocsThemeConfig = {
     toggleButton: true
   },
   toc: {
+    backToTop: true,
     extraContent: (
       // eslint-disable-next-line @next/next/no-img-element -- ignore since url is external and dynamic
       <img alt="placeholder cat" src="https://placekitten.com/g/300/200" />