+++ packages/nextra-theme-docs/src/components/flexsearch.tsx
@@ -1,6 +1,8 @@
 import cn from 'clsx'
+// flexsearch types are incorrect, they were overwritten in tsconfig.json
 import FlexSearch from 'flexsearch'
 import { useRouter } from 'next/router'
+import type { SearchData } from 'nextra'
 import type { ReactElement, ReactNode } from 'react'
 import { useCallback, useState } from 'react'
 import { DEFAULT_LOCALE } from '../constants'
@@ -37,13 +39,6 @@ type Result = {
   children: ReactNode
 }
 
-type NextraData = {
-  [route: string]: {
-    title: string
-    data: Record<string, string>
-  }
-}
-
 // This can be global for better caching.
 const indexes: {
   [locale: string]: [PageIndex, SectionIndex]
@@ -68,7 +63,7 @@ const loadIndexesImpl = async (
   const response = await fetch(
     `${basePath}/_next/static/chunks/nextra-data-${locale}.json`
   )
-  const data = (await response.json()) as NextraData
+  const searchData = (await response.json()) as SearchData
 
   const pageIndex: PageIndex = new FlexSearch.Document({
     cache: 100,
@@ -102,17 +97,16 @@ const loadIndexesImpl = async (
   })
 
   let pageId = 0
-  for (const route in data) {
+
+  for (const [route, structurizedData] of Object.entries(searchData)) {
     let pageContent = ''
     ++pageId
 
-    for (const heading in data[route].data) {
-      const [hash, text] = heading.split('#')
-      const url = route + (hash ? '#' + hash : '')
-      const title = text || data[route].title
-
-      const content = data[route].data[heading] || ''
-      const paragraphs = content.split('\n').filter(Boolean)
+    for (const [key, content] of Object.entries(structurizedData.data)) {
+      const [headingId, headingValue] = key.split('#')
+      const url = route + (headingId ? '#' + headingId : '')
+      const title = headingValue || structurizedData.title
+      const paragraphs = content.split('\n')
 
       sectionIndex.add({
         id: url,
@@ -139,7 +133,7 @@ const loadIndexesImpl = async (
 
     pageIndex.add({
       id: pageId,
-      title: data[route].title,
+      title: structurizedData.title,
       content: pageContent
     })
   }