+++ packages/nextra-theme-docs/src/mdx-components.tsx
@@ -15,6 +15,7 @@ function HeadingLink({
   context,
   children,
   id,
+  className,
   ...props
 }: ComponentProps<'h2'> & {
   tag: `h${2 | 3 | 4 | 5 | 6}`
@@ -45,25 +46,33 @@ function HeadingLink({
 
   return (
     <Tag
-      className={cn(
-        'nx-font-semibold nx-tracking-tight nx-text-slate-900 dark:nx-text-slate-100',
-        {
-          h2: 'nx-mt-10 nx-border-b nx-pb-1 nx-text-3xl nx-border-neutral-200/70 contrast-more:nx-border-neutral-400 dark:nx-border-primary-100/10 contrast-more:dark:nx-border-neutral-400',
-          h3: 'nx-mt-8 nx-text-2xl',
-          h4: 'nx-mt-8 nx-text-xl',
-          h5: 'nx-mt-8 nx-text-lg',
-          h6: 'nx-mt-8 nx-text-base'
-        }[Tag]
-      )}
+      className={
+        // can be added by footnotes
+        className === 'sr-only'
+          ? 'nx-sr-only'
+          : cn(
+              'nx-font-semibold nx-tracking-tight nx-text-slate-900 dark:nx-text-slate-100',
+              {
+                h2: 'nx-mt-10 nx-border-b nx-pb-1 nx-text-3xl nx-border-neutral-200/70 contrast-more:nx-border-neutral-400 dark:nx-border-primary-100/10 contrast-more:dark:nx-border-neutral-400',
+                h3: 'nx-mt-8 nx-text-2xl',
+                h4: 'nx-mt-8 nx-text-xl',
+                h5: 'nx-mt-8 nx-text-lg',
+                h6: 'nx-mt-8 nx-text-base'
+              }[Tag]
+            )
+      }
       {...props}
     >
       {children}
-      <span className="nx-absolute -nx-mt-20" id={id} ref={obRef} />
-      <a
-        href={`#${id}`}
-        className="subheading-anchor"
-        aria-label="Permalink for this section"
-      />
+      {id && (
+        <a
+          href={`#${id}`}
+          id={id}
+          className="subheading-anchor"
+          aria-label="Permalink for this section"
+          ref={obRef}
+        />
+      )}
     </Tag>
   )
 }
@@ -139,7 +148,7 @@ const Summary = (props: ComponentProps<'summary'>): ReactElement => {
     <summary
       className={cn(
         'nx-flex nx-items-center nx-cursor-pointer nx-list-none nx-p-1 nx-transition-colors hover:nx-bg-gray-100 dark:hover:nx-bg-neutral-800',
-        "before:nx-mr-1 before:nx-inline-block before:nx-transition-transform before:nx-content-[''] dark:before:nx-invert",
+        "before:nx-mr-1 before:nx-inline-block before:nx-transition-transform before:nx-content-[''] dark:before:nx-invert before:nx-shrink-0",
         'rtl:before:nx-rotate-180 [[data-expanded]>&]:before:nx-rotate-90'
       )}
       {...props}
@@ -215,7 +224,12 @@ export const getComponents = ({
         {...props}
       />
     ),
-    hr: props => <hr className="nx-my-8 dark:nx-border-gray-900" {...props} />,
+    hr: props => (
+      <hr
+        className="nx-my-8 nx-border-neutral-200/70 contrast-more:nx-border-neutral-400 dark:nx-border-primary-100/10 contrast-more:dark:nx-border-neutral-400"
+        {...props}
+      />
+    ),
     a: Link,
     table: props => (
       <Table