+++ packages/nextra/src/compile.ts
@@ -9,6 +9,7 @@ import grayMatter from 'gray-matter'
 import rehypeKatex from 'rehype-katex'
 import type { Options as RehypePrettyCodeOptions } from 'rehype-pretty-code'
 import rehypePrettyCode from 'rehype-pretty-code'
+import rehypeRaw from 'rehype-raw'
 import remarkGfm from 'remark-gfm'
 import remarkMath from 'remark-math'
 import remarkReadingTime from 'remark-reading-time'
@@ -17,6 +18,7 @@ import {
   CODE_BLOCK_FILENAME_REGEX,
   CWD,
   DEFAULT_LOCALE,
+  ERROR_ROUTES,
   MARKDOWN_URL_EXTENSION_REGEX
 } from './constants'
 import {
@@ -25,13 +27,19 @@ import {
   remarkCustomHeadingId,
   remarkHeadings,
   remarkLinkRewrite,
+  remarkMdxDisableExplicitJsx,
   remarkRemoveImports,
   remarkReplaceImports,
   remarkStaticImage,
-  structurize
+  remarkStructurize
 } from './mdx-plugins'
 import theme from './theme.json'
-import type { LoaderOptions, PageOpts, ReadingTime } from './types'
+import type {
+  LoaderOptions,
+  PageOpts,
+  ReadingTime,
+  StructurizedData
+} from './types'
 import { truthy } from './utils'
 
 globalThis.__nextra_temp_do_not_use = () => {
@@ -72,40 +80,60 @@ type MdxOptions = LoaderOptions['mdxOptions'] &
 // because we already use `remarkLinkRewrite` function to remove .mdx? extensions
 const clonedRemarkLinkRewrite = remarkLinkRewrite.bind(null)
 
+type CompileMdxOptions = Pick<
+  LoaderOptions,
+  | 'staticImage'
+  | 'flexsearch'
+  | 'defaultShowCopyCode'
+  | 'readingTime'
+  | 'latex'
+  | 'codeHighlight'
+> & {
+  mdxOptions?: MdxOptions
+  route?: string
+  locale?: string
+  filePath?: string
+  useCachedCompiler?: boolean
+  isPageImport?: boolean
+}
+
 export async function compileMdx(
   source: string,
-  loaderOptions: Pick<
-    LoaderOptions,
-    | 'staticImage'
-    | 'flexsearch'
-    | 'defaultShowCopyCode'
-    | 'readingTime'
-    | 'latex'
-    | 'codeHighlight'
-  > & { mdxOptions?: MdxOptions; route?: string; locale?: string } = {},
-  { filePath = '', useCachedCompiler = false, isPageImport = true } = {}
+  {
+    staticImage,
+    flexsearch,
+    readingTime,
+    latex,
+    codeHighlight,
+    defaultShowCopyCode,
+    route = '',
+    locale,
+    mdxOptions,
+    filePath = '',
+    useCachedCompiler,
+    isPageImport = true
+  }: CompileMdxOptions = {}
 ) {
   // Extract frontMatter information if it exists
   const { data: frontMatter, content } = grayMatter(source)
 
-  const structurizedData = Object.create(null)
-
   let searchIndexKey: string | null = null
-  if (typeof loaderOptions.flexsearch === 'object') {
-    if (loaderOptions.flexsearch.indexKey) {
-      searchIndexKey = loaderOptions.flexsearch.indexKey(
-        filePath,
-        loaderOptions.route || '',
-        loaderOptions.locale
-      )
+  if (
+    ERROR_ROUTES.has(route) ||
+    route === '/_app' /* remove this check in v3 */
+  ) {
+    /* skip */
+  } else if (typeof flexsearch === 'object') {
+    if (flexsearch.indexKey) {
+      searchIndexKey = flexsearch.indexKey(filePath, route, locale)
       if (searchIndexKey === '') {
-        searchIndexKey = loaderOptions.locale || DEFAULT_LOCALE
+        searchIndexKey = locale || DEFAULT_LOCALE
       }
     } else {
-      searchIndexKey = loaderOptions.locale || DEFAULT_LOCALE
+      searchIndexKey = locale || DEFAULT_LOCALE
     }
-  } else if (loaderOptions.flexsearch) {
-    searchIndexKey = loaderOptions.locale || DEFAULT_LOCALE
+  } else if (flexsearch) {
+    searchIndexKey = locale || DEFAULT_LOCALE
   }
 
   const {
@@ -116,7 +144,7 @@ export async function compileMdx(
     rehypePlugins,
     rehypePrettyCodeOptions
   }: MdxOptions = {
-    ...loaderOptions.mdxOptions,
+    ...mdxOptions,
     // You can override MDX options in the frontMatter too.
     ...frontMatter.mdxOptions
   }
@@ -124,22 +152,47 @@ export async function compileMdx(
   const format =
     _format === 'detect' ? (filePath.endsWith('.mdx') ? 'mdx' : 'md') : _format
 
-  const {
-    staticImage,
-    flexsearch,
-    readingTime,
-    latex,
-    codeHighlight,
-    defaultShowCopyCode
-  } = loaderOptions
-
   // https://github.com/shuding/nextra/issues/1303
   const isFileOutsideCWD =
     !isPageImport && path.relative(CWD, filePath).startsWith('..')
 
+  const isRemoteContent = outputFormat === 'function-body'
+
   const compiler =
-    (useCachedCompiler && cachedCompilerForFormat[format]) ||
-    (cachedCompilerForFormat[format] = createProcessor({
+    !useCachedCompiler || isRemoteContent
+      ? createCompiler()
+      : (cachedCompilerForFormat[format] ??= createCompiler())
+  const processor = compiler()
+
+  try {
+    const vFile = await processor.process(
+      filePath ? { value: content, path: filePath } : content
+    )
+
+    const { title, hasJsxInH1, readingTime, structurizedData } = vFile.data as {
+      readingTime?: ReadingTime
+      structurizedData: StructurizedData
+      title?: string
+    } & Pick<PageOpts, 'hasJsxInH1'>
+    // https://github.com/shuding/nextra/issues/1032
+    const result = String(vFile).replaceAll('__esModule', '_\\_esModule')
+
+    return {
+      result,
+      ...(title && { title }),
+      ...(hasJsxInH1 && { hasJsxInH1 }),
+      ...(readingTime && { readingTime }),
+      ...(searchIndexKey !== null && { searchIndexKey, structurizedData }),
+      ...(isRemoteContent && { headings: vFile.data.headings }),
+      frontMatter
+    }
+  } catch (err) {
+    console.error(`[nextra] Error compiling ${filePath}.`)
+    throw err
+  }
+
+  function createCompiler(): Processor {
+    return createProcessor({
       jsx,
       format,
       outputFormat,
@@ -157,12 +210,19 @@ export async function compileMdx(
             storageKey: 'selectedPackageManager'
           }
         ] satisfies Pluggable,
-        outputFormat === 'function-body' && remarkRemoveImports,
+        isRemoteContent && remarkRemoveImports,
         remarkGfm,
+        format !== 'md' &&
+          ([
+            remarkMdxDisableExplicitJsx,
+            // Replace the <summary> and <details> with customized components
+            { whiteList: ['details', 'summary'] }
+          ] satisfies Pluggable),
         remarkCustomHeadingId,
-        remarkHeadings,
+        [remarkHeadings, { isRemoteContent }] satisfies Pluggable,
+        // structurize should be before remarkHeadings because we attach #id attribute to heading node
+        flexsearch && ([remarkStructurize, flexsearch] satisfies Pluggable),
         staticImage && remarkStaticImage,
-        searchIndexKey !== null && structurize(structurizedData, flexsearch),
         readingTime && remarkReadingTime,
         latex && remarkMath,
         isFileOutsideCWD && remarkReplaceImports,
@@ -178,6 +238,12 @@ export async function compileMdx(
       ].filter(truthy),
       rehypePlugins: [
         ...(rehypePlugins || []),
+        format === 'md' && [
+          // To render <details /> and <summary /> correctly
+          rehypeRaw,
+          // fix Error: Cannot compile `mdxjsEsm` node for npm2yarn and mermaid
+          { passThrough: ['mdxjsEsm', 'mdxJsxFlowElement'] }
+        ],
         [parseMeta, { defaultShowCopyCode }],
         codeHighlight !== false &&
           ([
@@ -190,31 +256,6 @@ export async function compileMdx(
         attachMeta,
         latex && rehypeKatex
       ].filter(truthy)
-    }))
-
-  try {
-    compiler.data('headingMeta', { headings: [] })
-    const vFile = await compiler.process(
-      filePath ? { value: content, path: filePath } : content
-    )
-
-    const headingMeta = compiler.data('headingMeta') as Pick<
-      PageOpts,
-      'headings' | 'hasJsxInH1' | 'title'
-    >
-    const readingTime = vFile.data.readingTime as ReadingTime | undefined
-
-    return {
-      // https://github.com/shuding/nextra/issues/1032
-      result: String(vFile).replaceAll('__esModule', '_\\_esModule'),
-      ...headingMeta,
-      ...(readingTime && { readingTime }),
-      structurizedData,
-      searchIndexKey,
-      frontMatter
-    }
-  } catch (err) {
-    console.error(`[nextra] Error compiling ${filePath}.`)
-    throw err
+    })
   }
 }